guidance/requirements.txt
python-decouple==3.8
redis==5.0.4
SQLAlchemy==2.0.30
jupyter-events==0.6.3
jupyter-server==1.24.0
jupyter-ydoc==0.2.5
jupyter_client==7.4.9
jupyter_core==4.12.0
jupyter_server_fileid==0.9.1
jupyter_server_ydoc==0.8.0
jupyterlab==3.6.7
jupyterlab-pygments==0.2.2
jupyterlab_server==2.24.0